[preparation] Fix Jira SWBUG-7770: Champ occupation en %
Merged revision(s) 30211 from branches/sword-scipio-stab.

+++ src/libraries/clients_kernel/Architecture_ABC.h
@@ -41,7 +41,7 @@ public:
 
     //! @name Operations
     //@{
-    virtual void Initialize( const ObjectTypes& objectTypes, float height, unsigned int floorNumber, unsigned int parkingFloors, float occupation, 
+    virtual void Initialize( const ObjectTypes& objectTypes, unsigned int height, unsigned int floorNumber, unsigned int parkingFloors, float occupation, 
                              float trafficability, const std::string& material = "", const std::string& roofShape = "" ) = 0;
     virtual const MaterialCompositionType& GetMaterial() const = 0;
     virtual const RoofShapeType& GetRoofShape() const = 0;
