+++ src/libraries/gaming/Architecture.cpp
@@ -21,11 +21,11 @@ Architecture::Architecture( const sword::UrbanAttributes& message, kernel::Prope
     if( message.has_architecture() )
     {
         const sword::UrbanAttributes::Architecture& architecture = message.architecture();
-        Initialize( objectTypes, architecture.height(), architecture.floor_number(), architecture.parking_floors(), architecture.occupation(),
+        Initialize( objectTypes, static_cast< unsigned int >( architecture.height() ), architecture.floor_number(), architecture.parking_floors(), architecture.occupation(),
             architecture.trafficability(), architecture.material(), architecture.roof_shape() );
     }
     else
-        Initialize( objectTypes, 20.f, 6, 0, 0.5f, 0.5f );
+        Initialize( objectTypes, 20, 6, 0, 0.5f, 0.5f );
     CreateDictionnary( true );
 }
